chore: run command separate for each benchmark + refactor / clean up
