+++ .github/workflows/page-load-benchmark-pr.yml
@@ -40,7 +40,7 @@ jobs:
         run: yarn playwright install chromium
 
       - name: Run page load benchmarks
-        run: yarn test:e2e:benchmark
+        run: yarn test:e2e:benchmark:page-load
 
       - name: Upload benchmark results (github artifacts)
         uses: actions/upload-artifact@v4
